Scaffold shadcn-ui and tailwind-config packages
