+++ packages/design-system/package.json
@@ -34,6 +34,7 @@
     "@radix-ui/react-toggle-group": "^1.1.0",
     "@radix-ui/react-tooltip": "^1.1.2",
     "@repo/database": "workspace:*",
+    "@repo/tailwind-config": "workspace:*",
     "@segment/analytics-next": "^1.72.1",
     "@segment/analytics-node": "^2.1.2",
     "@sentry/nextjs": "^8.27.0",
