+++ packages/shadcn-ui/components.json
@@ -1,0 +1,20 @@
+{
+  "$schema": "https://ui.shadcn.com/schema.json",
+  "style": "new-york",
+  "rsc": true,
+  "tsx": true,
+  "tailwind": {
+    "config": "../tailwind-config/tailwind.config.ts",
+    "css": "../design-system/styles/globals.css",
+    "baseColor": "neutral",
+    "cssVariables": true,
+    "prefix": ""
+  },
+  "aliases": {
+    "components": "@repo/shadcn-ui/components",
+    "utils": "@repo/shadcn-ui/lib/utils",
+    "ui": "@repo/shadcn-ui/components/ui",
+    "lib": "@repo/shadcn-ui/lib",
+    "hooks": "@repo/shadcn-ui/hooks"
+  }
+}
