+++ packages/shadcn-ui/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "@repo/shadcn-ui",
+  "version": "0.0.0",
+  "private": true,
+  "peerDependencies": {
+    "react": "18.3.1"
+  },
+  "dependencies": {
+    "@repo/tailwind-config": "workspace:*",
+    "@radix-ui/react-icons": "^1.3.0",
+    "class-variance-authority": "^0.7.0",
+    "clsx": "^2.1.1",
+    "lucide-react": "^0.438.0",
+    "tailwind-merge": "^2.5.2",
+    "tailwindcss-animate": "^1.0.7"
+  },
+  "devDependencies": {
+    "@repo/typescript-config": "workspace:*",
+    "@types/node": "22.5.1",
+    "@types/react": "^18.3.4",
+    "typescript": "^5.5.4"
+  }
+}
