+++ packages/shadcn-ui/tailwind.config.ts
@@ -1,0 +1,1 @@
+export { default } from '@repo/tailwind-config';
